Move to module scalaparallelcollections
