--- core/js/src/main/scala/zio/prelude/PlatformSpecificDerive.scala
@@ -1,3 +1,0 @@
-package zio.prelude
-
-trait PlatformSpecificDerive {}
