--- core/js/src/main/scala/zio/prelude/PlatformSpecificEqualInstances.scala
@@ -1,3 +1,0 @@
-package zio.prelude
-
-trait PlatformSpecificEqualInstances {}
