--- core/js/src/main/scala/zio/prelude/PlatformSpecificInvariantInstances.scala
@@ -1,3 +1,0 @@
-package zio.prelude
-
-trait PlatformSpecificInvariantInstances {}
