--- core/js/src/main/scala/zio/prelude/PlatformSpecificOrdInstances.scala
@@ -1,3 +1,0 @@
-package zio.prelude
-
-trait PlatformSpecificOrdInstances {}
