--- core/js/src/main/scala/zio/prelude/PlatformSpecificPartialOrdInstances.scala
@@ -1,3 +1,0 @@
-package zio.prelude
-
-trait PlatformSpecificPartialOrdInstances {}
