--- core/native/src/main/scala/zio/prelude/PlatformSpecificSyntax.scala
@@ -1,3 +1,0 @@
-package zio.prelude
-
-trait PlatformSpecificSyntax {}
