+++ core/shared/src/main/scala/zio/prelude/Hash.scala
@@ -99,7 +99,7 @@ trait Hash[-A] extends Equal[A] { self =>
     )
 }
 
-object Hash extends Lawful[Hash] with PlatformSpecificHashInstances {
+object Hash extends Lawful[Hash] {
 
   /**
    * For all values `a1` and `a2`, if `a1` is equal to `a2` then the hash of
