+++ core/shared/src/main/scala/zio/prelude/package.scala
@@ -47,7 +47,6 @@ package object prelude
     with NonEmptyForEachSyntax
     with OrdSyntax
     with PartialOrdSyntax
-    with PlatformSpecificSyntax
     with ForEachSyntax
     with ZivariantSyntax {
