+++ scala-parallel-collections/src/main/scala/zio/prelude/scalaparallelcollections/package.scala
@@ -1,0 +1,11 @@
+package zio.prelude
+
+package object scalaparallelcollections
+    extends AssociativeFlattenInstances
+    with SPCDerive
+    with EqualInstances
+    with HashInstances
+    with InvariantInstances
+    with OrdInstances
+    with PartialOrdInstances
+    with Syntax {}
